requirements.txt
blinker==1.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cors==1.0.1
docopt==0.6.2
ecdsa==0.18.0
filelock==3.12.0
Flask==2.3.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
future==0.18.3
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.0.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jose==1.0.0
MarkupSafe==2.1.2
Naked==0.1.32
pipreqs==0.4.13
pyasn1==0.5.0
pycparser==2.21
pycryptodome==3.17
PyJWT==2.6.0
PySocks==1.7.1
python-jose==3.3.0
PyYAML==6.0
requests==2.28.2
requests-file==1.5.1
rsa==4.9
shellescape==3.8.1
six==1.16.0
SQLAlchemy==1.4.18
tldextract==3.4.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.3.0
yarg==0.1.9
zope.event==4.6
zope.interface==6.0
